Fix warning from the nightlies
git-svn-id: http://root.cern.ch/svn/root/branches/v5-24-00-patches@31872 27541ba8-7e3a-0410-8455-c3a389f83636

+++ net/xrootd/src/xrootd/src/XrdCrypto/XrdCryptosslAux.cc
@@ -615,7 +615,7 @@ int XrdCryptosslASN1toUTC(ASN1_TIME *tsn1)
    if (!gmtime_r(&now, &gtn)) return etime;
    //
    // Calculate correction
-   int tzcor = difftime(mktime(&ltn), mktime(&gtn));
+   int tzcor = (int) difftime(mktime(&ltn), mktime(&gtn));
    //
    // Apply correction
    etime += tzcor;
